implement Transport-Helper-class, code-formatting

+++ pocket-api/src/main/kotlin/com/uwefuchs/demo/kotlin/pocket/Responses.kt
@@ -3,5 +3,5 @@ package com.uwefuchs.demo.kotlin.pocket
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.uwefuchs.demo.kotlin.pocket.api.Item
 
-internal class RetrieveResponse(@JsonProperty("list") val items: Map<Int, Item>) {
+internal class RetrieveResponse(@JsonProperty("list") val items: Map<Long, Item>) {
 }
